Migrate finatra to explicit interpreters
