+++ server/finatra-server/finatra-server-cats/src/main/scala/sttp/tapir/server/finatra/cats/package.scala
@@ -2,8 +2,6 @@ package sttp.tapir.server.finatra
 
 import _root_.cats.effect.Effect
 import com.twitter.inject.Logging
-import io.catbird.util.Rerunnable
-import io.catbird.util.effect._
 import sttp.monad.MonadError
 import sttp.tapir.Endpoint
 import sttp.tapir.server.ServerEndpoint
@@ -12,23 +10,22 @@ import scala.reflect.ClassTag
 
 package object cats {
   implicit class RichFinatraCatsEndpoint[I, E, O](e: Endpoint[I, E, O, Any]) extends Logging {
-    def toRoute[F[_]](logic: I => F[Either[E, O]])(implicit serverOptions: FinatraServerOptions, eff: Effect[F]): FinatraRoute = {
-      e.serverLogic(logic).toRoute
-    }
+    @deprecated("Use FinatraCatsServerInterpreter.toRoute", since = "0.17.1")
+    def toRoute[F[_]](logic: I => F[Either[E, O]])(implicit serverOptions: FinatraServerOptions, eff: Effect[F]): FinatraRoute =
+      FinatraCatsServerInterpreter.toRoute(e)(logic)
 
+    @deprecated("Use FinatraCatsServerInterpreter.toRouteRecoverErrors", since = "0.17.1")
     def toRouteRecoverErrors[F[_]](logic: I => F[O])(implicit
         eIsThrowable: E <:< Throwable,
         eClassTag: ClassTag[E],
         eff: Effect[F]
-    ): FinatraRoute = {
-      e.serverLogicRecoverErrors(logic).toRoute
-    }
+    ): FinatraRoute = FinatraCatsServerInterpreter.toRouteRecoverErrors(e)(logic)
   }
 
   implicit class RichFinatraCatsServerEndpoint[I, E, O, F[_]](e: ServerEndpoint[I, E, O, Any, F]) extends Logging {
-    def toRoute(implicit serverOptions: FinatraServerOptions, eff: Effect[F]): FinatraRoute = {
-      new RichFinatraServerEndpoint(e.endpoint.serverLogic(i => eff.toIO(e.logic(new CatsMonadError)(i)).to[Rerunnable].run)).toRoute
-    }
+    @deprecated("Use FinatraServerInterpreter.toRoute", since = "0.17.1")
+    def toRoute(implicit serverOptions: FinatraServerOptions, eff: Effect[F]): FinatraRoute =
+      FinatraCatsServerInterpreter.toRoute(e)
 
     private class CatsMonadError(implicit F: Effect[F]) extends MonadError[F] {
       override def unit[T](t: T): F[T] = F.pure(t)
