+++ server/finatra-server/finatra-server-cats/src/test/scala/sttp/tapir/server/finatra/cats/FinatraCatsTestServerInterpreter.scala
@@ -3,15 +3,15 @@ package sttp.tapir.server.finatra.cats
 import cats.data.NonEmptyList
 import cats.effect.{ContextShift, IO, Resource, Timer}
 import sttp.tapir.Endpoint
-import sttp.tapir.server.finatra.{FinatraRoute, FinatraServerInterpreter, FinatraServerOptions}
-import sttp.tapir.server.tests.ServerInterpreter
+import sttp.tapir.server.finatra.{FinatraRoute, FinatraTestServerInterpreter, FinatraServerOptions}
+import sttp.tapir.server.tests.TestServerInterpreter
 import sttp.tapir.server.{DecodeFailureHandler, ServerDefaults, ServerEndpoint}
 import sttp.tapir.tests.Port
 
 import scala.concurrent.ExecutionContext
 import scala.reflect.ClassTag
 
-class FinatraServerCatsInterpreter extends ServerInterpreter[IO, Any, FinatraRoute] {
+class FinatraCatsTestServerInterpreter extends TestServerInterpreter[IO, Any, FinatraRoute] {
   implicit val ec: ExecutionContext = scala.concurrent.ExecutionContext.Implicits.global
   implicit val contextShift: ContextShift[IO] = IO.contextShift(ec)
   implicit val timer: Timer[IO] = IO.timer(ec)
@@ -22,12 +22,12 @@ class FinatraServerCatsInterpreter extends ServerInterpreter[IO, Any, FinatraRou
   ): FinatraRoute = {
     implicit val serverOptions: FinatraServerOptions =
       FinatraServerOptions.default.copy(decodeFailureHandler = decodeFailureHandler.getOrElse(ServerDefaults.decodeFailureHandler))
-    e.toRoute
+    FinatraCatsServerInterpreter.toRoute(e)
   }
 
   override def routeRecoverErrors[I, E <: Throwable, O](e: Endpoint[I, E, O, Any], fn: I => IO[O])(implicit
       eClassTag: ClassTag[E]
-  ): FinatraRoute = e.toRouteRecoverErrors(fn)
+  ): FinatraRoute = FinatraCatsServerInterpreter.toRouteRecoverErrors(e)(fn)
 
-  override def server(routes: NonEmptyList[FinatraRoute]): Resource[IO, Port] = FinatraServerInterpreter.server(routes)
+  override def server(routes: NonEmptyList[FinatraRoute]): Resource[IO, Port] = FinatraTestServerInterpreter.server(routes)
 }
